refactor: move LockCommand to ContentType folder
