+++ EMS/core-bundle/src/Command/ContentType/LockCommand.php
@@ -2,15 +2,17 @@
 
 declare(strict_types=1);
 
-namespace EMS\CoreBundle\Command;
+namespace EMS\CoreBundle\Command\ContentType;
 
 use EMS\CommonBundle\Elasticsearch\Document\Document;
 use EMS\CommonBundle\Elasticsearch\Document\DocumentInterface;
 use EMS\CommonBundle\Search\Search;
 use EMS\CommonBundle\Service\ElasticaService;
+use EMS\CoreBundle\Commands;
 use EMS\CoreBundle\Entity\ContentType;
 use EMS\CoreBundle\Repository\ContentTypeRepository;
 use EMS\CoreBundle\Repository\RevisionRepository;
+use EMS\Helpers\Standard\Json;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
@@ -20,6 +22,8 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 
 final class LockCommand extends Command
 {
+    protected static $defaultName = Commands::CONTENT_TYPE_LOCK;
+
     private string $by;
     private ContentType $contentType;
     private bool $force;
@@ -37,11 +41,11 @@ final class LockCommand extends Command
 
     public const RESULT_SUCCESS = 0;
 
-    public const name = 'ems:contenttype:lock';
-    protected static $defaultName = self::name;
-
-    public function __construct(private readonly ContentTypeRepository $contentTypeRepository, private readonly ElasticaService $elasticaService, private readonly RevisionRepository $revisionRepository)
-    {
+    public function __construct(
+        private readonly ContentTypeRepository $contentTypeRepository,
+        private readonly ElasticaService $elasticaService,
+        private readonly RevisionRepository $revisionRepository
+    ) {
         parent::__construct();
     }
 
@@ -55,7 +59,7 @@ final class LockCommand extends Command
             ->addOption(self::OPTION_USER, null, InputOption::VALUE_REQUIRED, 'lock username', 'EMS_COMMAND')
             ->addOption(self::OPTION_FORCE, null, InputOption::VALUE_NONE, 'do not check for already locked revisions')
             ->addOption(self::OPTION_IF_EMPTY, null, InputOption::VALUE_NONE, 'lock if there are no pending locks for the same user')
-            ->addOption(self::OPTION_OUUID, null, InputOption::VALUE_OPTIONAL, 'lock a specific ouuid', null)
+            ->addOption(self::OPTION_OUUID, null, InputOption::VALUE_OPTIONAL, 'lock a specific ouuid')
         ;
     }
 
@@ -87,16 +91,13 @@ final class LockCommand extends Command
 
         $by = $input->getOption(self::OPTION_USER);
         if (!\is_string($by)) {
-            throw new \RuntimeException('Unexpected user name');
+            throw new \RuntimeException('Unexpected username');
         }
         $this->by = $by;
 
         if (null !== $input->getOption(self::OPTION_QUERY)) {
             $this->query = \strval($input->getOption('query'));
-            \json_decode($this->query, true);
-            if (\json_last_error() > 0) {
-                throw new \RuntimeException(\sprintf('Invalid json query %s', $this->query));
-            }
+            Json::decode($this->query, 'Invalid json query');
         }
 
         $this->force = true === $input->getOption(self::OPTION_FORCE);
